Moved exceptions to namespaces where they where used.
