+++ src/CUITe/CUITe.csproj
@@ -278,11 +278,11 @@
     <Compile Include="Controls\WpfControls\WpfTreeItem.cs" />
     <Compile Include="Controls\WpfControls\WpfWindow.cs" />
     <Compile Include="DataManager.cs" />
-    <Compile Include="Exceptions\GenericException.cs" />
-    <Compile Include="Exceptions\InvalidTraversalException.cs" />
-    <Compile Include="Exceptions\InvalidSearchPropertyNamesException.cs" />
-    <Compile Include="Exceptions\InvalidSearchPropertiesFormatException.cs" />
-    <Compile Include="Exceptions\ResourceNotFoundException.cs" />
+    <Compile Include="Controls\GenericException.cs" />
+    <Compile Include="Controls\InvalidTraversalException.cs" />
+    <Compile Include="Controls\InvalidSearchPropertyNamesException.cs" />
+    <Compile Include="SearchConfigurations\InvalidSearchPropertiesFormatException.cs" />
+    <Compile Include="ResourceNotFoundException.cs" />
     <Compile Include="ObjectRepositoryManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SearchConfigurations\By.cs" />
