+++ src/CUITe/Controls/GenericException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace CUITe
+namespace CUITe.Controls
 {
     public class GenericException : Exception
     {
